+++ src/components/BottomDots.vue
@@ -122,7 +122,7 @@
   </div>
 </template>
 
-<script lang="ts">
+<script>
 
 export default {
   components: {