+++ webpack.config.js
@@ -1,0 +1,10 @@
+// Webpack
+const path = require('path');
+
+module.exports = {
+    resolve: {
+    alias: {
+        '@': path.resolve(__dirname, "src"),
+        },
+    },
+};